+++ experimental/packages/opentelemetry-instrumentation-http/test/utils/DummyPropagation.ts
@@ -2,17 +2,28 @@
  * Copyright The OpenTelemetry Authors
  * SPDX-License-Identifier: Apache-2.0
  */
-import type { Context, TextMapPropagator } from '@opentelemetry/api';
+import type {
+  Context,
+  TextMapGetter,
+  TextMapPropagator,
+} from '@opentelemetry/api';
 import { trace, TraceFlags } from '@opentelemetry/api';
 import type * as http from 'http';
 
 export class DummyPropagation implements TextMapPropagator {
   static TRACE_CONTEXT_KEY = 'x-dummy-trace-id';
   static SPAN_CONTEXT_KEY = 'x-dummy-span-id';
-  extract(context: Context, carrier: http.OutgoingHttpHeaders) {
+  extract(
+    context: Context,
+    carrier: http.IncomingMessage,
+    getter: TextMapGetter<http.IncomingMessage>
+  ) {
     const extractedSpanContext = {
-      traceId: carrier[DummyPropagation.TRACE_CONTEXT_KEY] as string,
-      spanId: carrier[DummyPropagation.SPAN_CONTEXT_KEY] as string,
+      traceId: getter.get(
+        carrier,
+        DummyPropagation.TRACE_CONTEXT_KEY
+      ) as string,
+      spanId: getter.get(carrier, DummyPropagation.SPAN_CONTEXT_KEY) as string,
       traceFlags: TraceFlags.SAMPLED,
       isRemote: true,
     };